requirements/backend/livechat/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg==3.1.19
PyJWT
python-dotenv
pytz
sqlparse
typing_extensions
httpx
channels==4
daphne==4
